fix: Fix bug that number of elements in page was returned as total elements.

+++ backend/src/main/java/com/go_exchange_easier/backend/chat/room/impl/RoomServiceImpl.java
@@ -62,7 +62,7 @@ public class RoomServiceImpl implements RoomService {
                             new AuthorSummary(lastMessageAuthorNick,
                                     lastMessageAuthorAvatarKey))));
         }
-        return SimplePage.of(rooms, page, size, totalElements);
+        return SimplePage.of(rooms, page, size, (long) totalElements);
     }
 
     @Override
